Remove all parser and replace it by the FastParse implementation.
