+++ core/shared/src/main/scala/cron4s/errors.scala
@@ -1,17 +1,11 @@
 package cron4s
 
-import scala.util.parsing.input.Position
-import scalaz.NonEmptyList
 
 /**
   * Created by alonsodomin on 30/08/2016.
   */
-sealed trait ExpressionError
+sealed trait ParseFailed
 
-final case class ParseFailed(message: String, input: String, position: Position) extends ExpressionError {
-  override def toString = message + ":\n\n" + position.longString
-}
-
-final case class InvalidCronExpr(invalidFields: NonEmptyList[InvalidExpr]) extends ExpressionError
-final case class InvalidExpr(field: CronField, msg: String)
-final case class InvalidFieldExpr[F <: CronField](field: F, msg: String) extends ExpressionError
+final case class InvalidCron(msg: String) extends ParseFailed
+final case class InvalidField(field: CronField, msg: String)
+final case class InvalidFieldExpr[F <: CronField](field: F, msg: String) extends ParseFailed
